Add connection implementation.
Connection is a category which is defined in apq/connection.h
To obtain connection ConnectionProvider must exists. ConnectionProvider
category is defined in apq/connection.h too. Connection is
ConnectionProvider itself. To obtain a connection from connection
provider get_connection() function exists.
Connectable is a category of objects which contains Connection.
Connection is Connectable itself. Wrapped connection is Connectable
if wrapper is ConnectionWrapper.

The basic implementation of the Connection category can be found
in apq/impl/connection.h.

Calss connection_info is a ConnectionProvider which creates connection
via connection info string. To perform the connection async_connect
operation is defined.