Remove redundant asio::post() in operations
post() can not be cancelled by timer, so we need to remove it.
Since all the operations are using handler's executor there is
no need to do post() for properly dispatching. That's why it is
safe to remove it.

Also removed retry for the send_query_params() which wraps
PQSendQueryParams(). It looks like I understood the PostgreSQL
documentation wrong and this function does not need to retry
on error with non-blocking query send.