Make pooled_connection a Nullable
pooled_connection is a wrapper for the Connection concept
objects which is placed into a connection pool. It can be
in null state. So this commit adds necessary things about
to be conform to Nullable concept. pooled_connection is in
null state if the underlying handle is empty or the underlying
connection object is in null state.

The second thing is pooled_connection is using unwrap_impl
instead of unwrap_connection_impl to eliminate unwrap and
unwrap_connection implementations dualism.
Since pooled_connection is a wrapper the unwrap operation
should be implemented and unwrap_connection should use it
istead of own implementation.