[INCOMPATIBLE] Huge refactoring to decouple unwrap() and Nullable
RATIONALE

It seems like unwrap() and Nullable should be completely
decoupled. And unwrap() should be specialized for every
adapted type independently. So the functional objects are
related to unwrap() implementations should be easy accessable
and reuseable. That's why the reorganization is needed.

CHANGES

Changes default behavior of unwrap(). Can break user code.

Now by default the function implementation does not look if
the argument type is nullable and has dereference operator,
thus it does not try to dereference the nullable argument
by default. Now it just forward it until unwrap_impl template
is specialized for the argument type.

This commit may break user code in case of using custom
nullable type without unwrap_impl specialization. But it
gives more predictable behavior of unwrap() and completely
decouples unwrap() and Nullable as it should be.

If your code is broken after this commit - please specify
ozo::unwrap_impl for your type, as it was made for nullable
types in ext/std or ext/boost files.