Using recursive algorithms for connection functions
Since the functions unwrap connection object recusively,
null state should be determined the same recursive way.
Change default unwrap connection strategy to uwrap_recursive().

Previously recursive call of unwrap_connection was performed
for Nullable types only. But this is not the right way. The
right way is to call unwrap_connection if unwrapped type is
different to the type. This commit solves this problem.