Add ozo::none_t and ozo::none
RATIONALE

Sometimes it needs to point something as non existent or void.
The void type itself is not an ordinary type and can not have an object.
So thats why the ozo::none_t and ozo::none have been introduced.

ozo::none can substitute detail::do_nothing functional class.

Remove unused do_nothing functor as it is not needed anymore.