Fallback micro framework
Introduce fallback mechanism with retry strategy as a first implemented
strategy.

RATIONALE

Distributed environment is unstable, so we want to have recovery mechanisms.
Fallback microframework is created to provide such customizeable and easy
to use solution. Here we have two new concepts: Strategy and Try.

Strategy is entity which determines how to try to recover an operation which
failed. Try is an entity which holds a context of an operation and decides
should we try to recover a failed operation or give an error for a user.

OPERATION SUPPORT FALLBACK

 * request
 * execute
 * get_connection
 * begin

BASIC USAGE

auto conn = execute[fallback::retry(connection_error)*3](conn_info[io], query, 3s, yield[ec]);