Add deadline_handler to avoid use same timer for defferent operations
MOTIVATION
==========

Currently the library uses Connection-attached timer for time constrained
operations. This approach has disadvantages in terms of sharing single
timer object for different operations. Thus operation-attached timer is
the right solution. At first this approach should be used in the cancel()
implementation due to Connection-attached timer can not be used to time
constrain cancel() operation since it may be used during the different
operation is in progress.