Move to connection-bound executor
Since Boost 1.70 boost::asio::posix_descriptor does not provide
io_context() which was used as connection-bound io_context it has been
decided to move to connection-bound executor and remove direct usage of
io_context. It has a reference on connection, but removed from the
interfaces to avoid usage of deprecated get_io_context() methods of
boost::asio entities.

The commit is huge, but it is better now than later. Without this commit
no chance to remove connection-bound timer.