Reimplement transaction as Connection model
Transaction models Connection with all the necessary functions.
Transaction inherits copyable and nullable traits of the underlying
connection. It describes the open transaction on the underlying
connection. The transaction is not an RAII object it does not control
transaction flow. It is dedicated to describe current transaction
and represent a transaction on the type level.