Explicit pooled_connection
Problem

Connection in the pool is stored with a reference to
last used io_context and active IO stream that attached to.
The connection pool is io_context agnostic and is not served
by any particular io_context. So it can be a dangling reference
on io_context inside stored connection if the io_context object
has been destroyed.

The solution

The only right solution is to exclude `io_context` related entities
from stored in the pool objects. Since there is a difference in the
behavior of `connection` and `pooled_connection` it should be two
different types.