Update ozo::conection and Connection concept
Add `is_open()` requirement as it is needed to determine the
situation if the connection does not own the native handle.

Add `connection::release()` as it is needed to transfer the
connection handle to different implementations. E.g., it is
needed to implement a true io-context-unbound connection pool.

Using of `dup()` to make a duplicate of the connection file
descriptor is removed in favor of `posix_descriptor::release()`.
It should reduce the cost of the stream object attaching.

Update connection requirements and types

The `pooled_connection` should not provide write access to oid_map,
or allow to assign different connection handle. That's why some
functionality has been removed from the `Connection` concept
requirements.